Refactor protocols
